Conflicts:
	CMakeLists.txt
	plugins/src/texture/CMakeLists.txt

+++ artiste/src/vsx_widget/dialogs/vsx_widget_window_statics.cpp
@@ -1,6 +1,25 @@
+/**
+* Project: VSXu: Realtime visual programming language, music/audio visualizer, animation tool and much much more.
+*
+* @author Jonatan Wallmander, Robert Wenzel, Vovoid Media Technologies Copyright (C) 2003-2011
+* @see The GNU Public License (GPL)
+*
+* This program is free software; you can redistribute it and/or modify
+* it under the terms of the GNU General Public License as published by
+* the Free Software Foundation; either version 2 of the License, or
+* (at your option) any later version.
+*
+* This program is distributed in the hope that it will be useful, but
+* WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
+* or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License
+* for more details.
+*
+* You should have received a copy of the GNU General Public License along
+* with this program; if not, write to the Free Software Foundation, Inc.,
+* 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
+*/
+
 #ifndef VSX_NO_CLIENT
-//#include <string>
-//#include <iostream>
 #include <map>
 #include <list>
 #include "vsx_gl_global.h"
@@ -166,7 +185,7 @@ dialog_messagebox::dialog_messagebox(vsx_string title_,vsx_string hint) {
   }
   vsx_widget::init_children();
 
-  int max_len = 0;
+  size_t max_len = 0;
   for (unsigned long i = 0; i < lines.size(); ++i) {
     if (lines[i].size() > max_len) max_len = lines[i].size();
   }
