+++ artiste/src/vsx_widget/module_choosers/vsx_widget_module_chooser_list.cpp
@@ -1,3 +1,24 @@
+/**
+* Project: VSXu: Realtime visual programming language, music/audio visualizer, animation tool and much much more.
+*
+* @author Jonatan Wallmander, Robert Wenzel, Vovoid Media Technologies Copyright (C) 2003-2011
+* @see The GNU Public License (GPL)
+*
+* This program is free software; you can redistribute it and/or modify
+* it under the terms of the GNU General Public License as published by
+* the Free Software Foundation; either version 2 of the License, or
+* (at your option) any later version.
+*
+* This program is distributed in the hope that it will be useful, but
+* WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
+* or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License
+* for more details.
+*
+* You should have received a copy of the GNU General Public License along
+* with this program; if not, write to the Free Software Foundation, Inc.,
+* 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
+*/
+
 #ifndef VSX_NO_CLIENT
 #include <vector>
 #include <map>
@@ -89,7 +110,7 @@ public:
 		int prev_carety = editor->carety;
 		int prev_selected_line = editor->selected_line;
 		editor->event_mouse_down(distance, coords,0);  // ugly hack! LAMORZ ATTACK!
-		if (editor->selected_line < i_rows_lookup.size())
+		if ((size_t)editor->selected_line < i_rows_lookup.size())
 		{
 			mod_i = i_rows_lookup[editor->selected_line];
 		} else
