+++ engine/include/vsx_array.h
@@ -1,3 +1,24 @@
+/**
+* Project: VSXu: Realtime visual programming language, music/audio visualizer, animation tool and much much more.
+*
+* @author Jonatan Wallmander, Robert Wenzel, Vovoid Media Technologies Copyright (C) 2003-2011
+* @see The GNU Public License (GPL)
+*
+* This program is free software; you can redistribute it and/or modify
+* it under the terms of the GNU General Public License as published by
+* the Free Software Foundation; either version 2 of the License, or
+* (at your option) any later version.
+*
+* This program is distributed in the hope that it will be useful, but
+* WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
+* or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License
+* for more details.
+*
+* You should have received a copy of the GNU General Public License along
+* with this program; if not, write to the Free Software Foundation, Inc.,
+* 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
+*/
+
 #ifndef VSX_ARRAY_H
 #define VSX_ARRAY_H
 
@@ -15,14 +36,14 @@
 
 template<class T>
 class vsx_array {
-  unsigned long allocated;
-  unsigned long used;
+  size_t allocated;
+  size_t used;
   T* A;
-  unsigned long allocation_increment;
-  unsigned long data_volatile;
+  size_t allocation_increment;
+  size_t data_volatile;
 
 public:
-  unsigned long timestamp;
+  size_t timestamp;
 
   void set_allocation_increment(unsigned long new_increment) {
   	allocation_increment = new_increment;
@@ -64,18 +85,18 @@ public:
   T* get_end_pointer() {
     return &A[used-1];
   }
-  unsigned long get_allocated() {
+  size_t get_allocated() {
     return allocated;
   }
-  unsigned long get_used() {
+  size_t get_used() {
     return used;
   }
   // std::vector compatibility
-  unsigned long push_back(T val) {
+  size_t push_back(T val) {
     (*this)[used] = val;
     return used;
   }
-  unsigned long size() {
+  size_t size() {
     return used;
   }
 
@@ -131,28 +152,28 @@ public:
 
 template<class T>
 class vsx_avector_nd {
-  unsigned long allocated;
-  unsigned long used;
+  size_t allocated;
+  size_t used;
   T* A;
-  unsigned long allocation_increment;
+  size_t allocation_increment;
 public:
-  unsigned long timestamp;
+  size_t timestamp;
   T* get_pointer() {
     return A;
   }
-  unsigned long get_allocated() {
+  size_t get_allocated() {
     return allocated;
   }
-  unsigned long get_used() {
+  size_t get_used() {
     return used;
   }
   // std::vector compatibility
-  unsigned long push_back(T val) {
+  size_t push_back(T val) {
     (*this)[used] = val;
     return used;
   }
 
-  unsigned long size() {
+  size_t size() {
     return used;
   }
